Page object moved to source roots.
