--- singleGUIAutomationsProject/src/test/java/PageObjects/LandingPage.java
@@ -1,10 +1,0 @@
-package PageObjects;
-
-import net.serenitybdd.core.pages.PageObjects;
-import org.openqa.selenium.WebDriver;
-
-public class LandingPage extends PageObjects {
-    public LandingPage(WebDriver driver) {
-        super(driver);
-    }
-}
