+++ src/main/java/LandingPage.java
@@ -1,0 +1,10 @@
+package PageObjects;
+
+import net.serenitybdd.core.pages.PageObjects;
+import org.openqa.selenium.WebDriver;
+
+public class LandingPage extends PageObjects {
+    public LandingPage(WebDriver driver) {
+        super(driver);
+    }
+}
